Programs are now coinductive, more improvements
Various changes related to the addition of nontermination to the tarpit:

1. Programs are colists of instructions.

2. Made Free coinductive and renamed to EffectTree.

3. HS_IO is now a coinductive type closed over all the possible IO
actions that we need.

3.* Added a new constructor, HS_id, which just satisfies the
productivity requirement for when we don't need to actually execute any
action (for instance, when we are only changing the tape state); this of
course is extracted to Prelude.id.

4. The example program is basically a REPL.

Other unrelated improvements:

1. Replaced the (Plus n) instruction with Dec | Inc.
2. Improved extraction for numbers.

Finally, I now use my own fork of Coq to get around an annoying syntax
error that its Haskell extraction was creating. It's still possible to
use vanilla Coq, but you'll have to fix the import of GHC.Base every
time you extract.